Add usersprog

+++ kernel/definitions.h
@@ -2,7 +2,7 @@
 
 // uart.c
 void uart_init();
-void uart_send(unsigned int c);
+void uart_send(uint8_t c);
 char uart_getc();
 void uart_puts(uint8_t *s);
